+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 
 ## Unreleased
 
+* Add automatic wait-and-timeout to Toolproof actions that get elements
+
 ## v0.8.0 (December 3, 2024)
 
 * Add instructions for clicking and hovering elements on a page